- changed @ControllerAdvice to @RestControllerAdvice and dropeed @ResponseBody - changed all Exception handlers to return an ErrorDetails instead of a String - re-arranged ErrorDetails field so error and detail appear first - renamed BookNotFoundAdvice to BookControllerAdvice - fix wrong @ResponseStatue at UserExceptionAdvice#notFoundHandler
