--- projectskeleton/src/main/java/org/regeneration/exceptions/BookNotFoundAdvice.java
@@ -1,32 +1,0 @@
-package org.regeneration.exceptions;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-import java.util.Date;
-
-@ControllerAdvice
-public class BookNotFoundAdvice {
-
-    @ResponseBody
-    @ExceptionHandler(BookNotFoundException.class)
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    String bookNotFoundHandler(BookNotFoundException ex) {
-        return ex.getMessage();
-    }
-
-    @ResponseBody
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    ErrorDetails bookNotFoundHandler(MethodArgumentNotValidException ex) {
-        ErrorDetails errorDetails = new ErrorDetails(new Date(), "Validation Failed",
-                ex.getBindingResult().toString());
-        return errorDetails;
-    }
-
-
-}
